refactor: first pass move registry query code into environmemnt
change name of func
move away from bool return to existing error methodology
treat error and registy value as separate return values
add trace
add debug logging
add mocked version
add "not implemented" version
