+++ src/environment_unix.go
@@ -59,3 +59,7 @@ func (env *environment) getCachePath() string {
 	}
 	return env.homeDir()
 }
+
+func (env *environment) getWindowsRegistryKeyValue(regPath string, regKey string) (string, error) {
+	return "", errors.New("not implemented")
+}
