+++ src/environment_windows.go
@@ -3,9 +3,13 @@
 package main
 
 import (
+	"errors"
+	"fmt"
 	"os"
+	"strings"
 	"syscall"
 	"time"
+	"unsafe"
 
 	"github.com/Azure/go-ansiterm/winterm"
 	"golang.org/x/sys/windows"
@@ -98,4 +102,94 @@ func (env *environment) getCachePath() string {
 		return cachePath
 	}
 	return env.homeDir()
+}
+
+//
+// Takes a registry path like "HKLM\Software\Microsoft\Windows NT\CurrentVersion" and a key under that path like "CurrentVersion" (or "" if the (Default) key is required).
+// Returns a bool and string:
+//
+//   true and the retrieved value formatted into a string if successful.
+//   false and the string will be the error
+//
+func (env *environment) getWindowsRegistryKeyValue(regPath string, regKey string) (string, error) {
+	env.trace(time.Now(), "getRegistryKeyValue", regPath, regKey)
+
+	// Extract root HK value and turn it into a windows.Handle to open the key.
+	regPathParts := strings.SplitN(regPath, "\\", 2)
+
+	regRootHKeyHandle := getHKEYHandleFromAbbrString(regPathParts[0])
+	if regRootHKeyHandle == 0 {
+		errorLogMsg := fmt.Sprintf("Error, Supplied root HKEY value not valid: '%s'", regPathParts[0])
+		env.log(Error, "getRegistryKeyValue", errorLogMsg)
+		return "", errors.New(errorLogMsg)
+	}
+
+	// Second part of split is registry path after HK part - needs to be UTF16 to pass to the windows. API
+	regPathUTF16, regPathUTF16ConversionErr := windows.UTF16FromString(regPathParts[1])
+	if regPathUTF16ConversionErr != nil {
+		errorLogMsg := fmt.Sprintf("Error, Could not convert supplied path '%s' to UTF16, error: '%s'", regPathParts[1], regPathUTF16ConversionErr)
+		env.log(Error, "getRegistryKeyValue", errorLogMsg)
+		return "", errors.New(errorLogMsg)
+	}
+
+	// Ok - open it..
+	var hKeyHandle windows.Handle
+	regOpenErr := windows.RegOpenKeyEx(regRootHKeyHandle, &regPathUTF16[0], 0, windows.KEY_READ, &hKeyHandle)
+	if regOpenErr != nil {
+		errorLogMsg := fmt.Sprintf("Error RegOpenKeyEx opening registry path to '%s', error: '%s'", regPath, regOpenErr)
+		env.log(Error, "getRegistryKeyValue", errorLogMsg)
+		return "", errors.New(errorLogMsg)
+	}
+	// Success - from here on out, when returning make sure to close that reg key with a deferred call to close:
+	defer windows.RegCloseKey(hKeyHandle)
+
+	// Again - need UTF16 of the key for the API:
+	regKeyUTF16, regKeyUTF16ConversionErr := windows.UTF16FromString(regKey)
+	if regKeyUTF16ConversionErr != nil {
+		errorLogMsg := fmt.Sprintf("Error, could not convert supplied key '%s' to UTF16, error: '%s'", regKey, regKeyUTF16ConversionErr)
+		env.log(Error, "getRegistryKeyValue", errorLogMsg)
+		return "", errors.New(errorLogMsg)
+	}
+
+	// Two stage way to get the key value - query once to get size - then allocate and query again to fill it.
+	var keyBufType uint32
+	var keyBufSize uint32
+
+	regQueryErr := windows.RegQueryValueEx(hKeyHandle, &regKeyUTF16[0], nil, &keyBufType, nil, &keyBufSize)
+	if regQueryErr != nil {
+		errorLogMsg := fmt.Sprintf("Error calling RegQueryValueEx to retrieve key data size with error '%s'", regQueryErr)
+		env.log(Error, "getRegistryKeyValue", errorLogMsg)
+		return "", errors.New(errorLogMsg)
+	}
+
+	// Alloc and fill...
+	var keyBuf = make([]byte, keyBufSize, keyBufSize)
+
+	regQueryErr = windows.RegQueryValueEx(hKeyHandle, &regKeyUTF16[0], nil, &keyBufType, &keyBuf[0], &keyBufSize)
+	if regQueryErr != nil {
+		errorLogMsg := fmt.Sprintf("Error calling RegQueryValueEx to retrieve key data with error '%s'", regQueryErr)
+		env.log(Error, "getRegistryKeyValue", errorLogMsg)
+		return "", errors.New(errorLogMsg)
+	}
+
+	// Format result into a string, depending on type.  (future refactor - move this out into it's own function)
+	switch keyBufType {
+	case windows.REG_SZ:
+		var uint16p *uint16
+		uint16p = (*uint16)(unsafe.Pointer(&keyBuf[0])) // nasty casty
+
+		valueString := windows.UTF16PtrToString(uint16p)
+		env.log(Debug, "getRegistryKeyValue", fmt.Sprintf("success, string: %s", valueString))
+		return valueString, nil
+	case windows.REG_DWORD:
+		var uint32p *uint32
+		uint32p = (*uint32)(unsafe.Pointer(&keyBuf[0])) // more casting goodness
+
+		valueString := fmt.Sprintf("0x%08X", *uint32p)
+		env.log(Debug, "getRegistryKeyValue", fmt.Sprintf("success, DWORD, formatted as string: %s", valueString))
+		return valueString, nil
+	default:
+		errorLogMsg := fmt.Sprintf("Error, no formatter for REG_? type:%d, data size:%d bytes", keyBufType, keyBufSize)
+		return "", errors.New(errorLogMsg)
+	}
 }
